+++ thotlib/dialogue/AmayaApplyClassPanel.cpp
@@ -124,7 +124,7 @@ void AmayaApplyClassPanel::RefreshToolTips()
  */
 void AmayaApplyClassPanel::SendDataToPanel( AmayaParams& p )
 {
-  int nb_class              = (int)p.param1;
+  int nb_class              = (long int)p.param1;
   const char * listBuffer   = (char *)p.param2;
   const char * currentClass = (char *)p.param3;
   int ref                   = (int)p.param4;;
