+++ thotlib/dialogue/AmayaMathMLPanel.cpp
@@ -143,7 +143,7 @@ void AmayaMathMLPanel::OnButton( wxCommandEvent& event )
  */
 void AmayaMathMLPanel::SendDataToPanel( AmayaParams& p )
 {
-  int action = (int)p.param1;
+  int action = (long int)p.param1;
   if (action == wxMATHML_ACTION_INIT)
     {
     }
